Fixing clang-format error

+++ libs/hashing/include/hpx/hashing/fibhash.hpp
@@ -12,14 +12,14 @@
 
 #include <hpx/config.hpp>
 
-#include <cstddef>
+#include <cstdint>
 #include <cstdlib>
 
 namespace hpx { namespace util {
 
     namespace detail {
 
-        template <std::size_t N>
+        template <std::uint64_t N>
         struct hash_helper;
 
         template <>
@@ -28,7 +28,7 @@ namespace hpx { namespace util {
             HPX_STATIC_CONSTEXPR int log2 = -1;
         };
 
-        template <std::size_t N>
+        template <std::uint64_t N>
         struct hash_helper
         {
             HPX_STATIC_CONSTEXPR std::uint64_t log2 =
@@ -36,13 +36,14 @@ namespace hpx { namespace util {
             HPX_STATIC_CONSTEXPR std::uint64_t shift_amount = 64 - log2;
         };
 
-        HPX_STATIC_CONSTEXPR std::uint64_t golden_ratio = 11400714819323198485llu;
+        HPX_STATIC_CONSTEXPR std::uint64_t golden_ratio =
+            11400714819323198485llu;
     }    // namespace detail
 
     // This function calculates the hash based on a multiplicative Fibonacci
     // scheme
-    template <std::size_t N>
-    HPX_CONSTEXPR std::size_t fibhash(std::uint64_t i)
+    template <std::uint64_t N>
+    HPX_CONSTEXPR std::uint64_t fibhash(std::uint64_t i)
     {
         using helper = detail::hash_helper<N>;
         static_assert(N != 0, "This algorithm only works with N != 0");
